Fixed DistributionTest
-It looks that the problem had place in the 'channel' of the JGroups
implementations, because it was never close but in both test's it had
the same 'namegroup'. So individually, both test's work correctly, but
when executing in the same execution, the channel was already in use
and did not let others to send or receive (i'm not sure) messages
-There is a warning that might let into problems, but for now it have
not been relevant in the execution
-he main solution was because in the
involved classes it was added a close (channel) service, also it was
implemented.
-In the classJGroupsAbstractFacade it was changed some configurations of
the UDP so it wont appear more warnings in communication.