Fixed generated problems
As explained in the others commit, this implementation has to be
modular, so the generated classes must be un-editable. Therefore, there
must be classes that extend from the generated classes to avoid problems
when the time comes to re-generate the mw2e artifacts.
-Created a EketalMavenWizard, and the plugin.xml will point to this
class.
-Edited the MavenProjectCreator Class: Adding the folders of the maven
style, and the maven nature
-EketalNewProjectWizardInitialContents: Change to generate the default
class in a source folder with the maven style
This implementation tries to create the pom, because the maven nature is
already added, but not working as desire