Migrate to JBase solved #18
The aspectj class needs to work with Java (not xtend) code. So the code
inside the reaction must be normal java code. For this purpose, there is
a tool named JBase, and this commit approach this change.
The change in the Generator accomplish the use of the JBase migration,
adding methods and using the imports from the language.