Fixed the reflexive communication with JGroups
First it was changed the Event definition from the Kernel and some
classes for communication, and the NamedEvent class had uncompleted
methods. Finally, the DeadLock example added an main class to run the
program, and the jar of the kernel in the lib.osgi project.
The Inferrer added a close method for the eventbrokers to finished the
channels that were opened