Datarace example
This commit achieves the JBossCache example, the main changes are: 
pom.xml: it was toggled the jar creator.
README.md's: Synchronized the README files with the Datarace example. In
the parent's README was changed the Deadlock example by the Datarace
example.
Datarace.eketal: Changed the automaton definition.
RunExecute and StartExecute: stops communication of the channel created
by Ketal and how they print the messages.