JBossCache with JGroups 2.6.13
This commit achieves the correct work of the program without a mock. So
when executing the Datarace example, it starts a Channel for JBossCache
and other two for the Eketal library.
First, the total-replication archive was put into the src.main.java
maven directory, this is to be found in the jar.
The lib jar from the lib.osgi project was updated
The pom from the lib project was changed, in particular, the version of
the commons.loggin dependency
Later, it was changed the log4j dependency by the commons.loggin
Finally, the logger of the RunExecute and the StartExecute class was
changed for the ones in the commons.loggin