Log4J implementation
First, it was changed the Log4J version on the lib project for desempeño
purposes, this new implementation must changes the log4j-1x scheme to
the next generation (log4j-2x). And all the classes that use it must be
updated.
Also, the generator and inferrer had to be updated to use the same log4j
version, that was the last.