Improvement of the language
This commit achieves a new support for the language, this for the Hadoop
example purpose.
First, (in the xtext file) the main change in the language that is in
the Event Declaration of the pointcut,i was obligatory to put a kind of
return, however, the interceptions of the constructors does not return
anything, so, the return kind has to be optional.
Now, the generator was updated to understand when the intercept has no
return.
Finally, it was added a validation for avoiding repeated event's names
definitions.