Multi-automatons improvement
First, it was changed the name of the classes to always start with the
'_' character, this for avoiding repeated class names.
Second, In the Inferrer and the Generaor class it was changed the
Reaction, the EventHandler, the Aspectj classes to support multiple
automatons in the pointcuts.
Third, As the names of somes classes were chenged, this changed must be
reflexed in all the other classes that had a references of the mentioned
before.