Fix Exception launched by the split method and cleaned the project
The split method interpreted the ":[" as a regular expression, so it was
launched a pattern exception in the Inferrer class
There was a clean of the unused imports (Inferrer, Validator and
generator).
The Test class of the Deadlock was updated with the name of the events
(the test failed because of incorrect event names).
The .eketal class was updated with the host tag
The executor changed one line to review the current state of the
automaton