Port to xtext 2.12.0 close #27
This project was port to xtext 2.12.0
Artifacts that need to be updated are the master pom (with correct
versions of the upgrade), MANIFEST from the language definition project,
because there is a bundle that need to be updated. Finally, the
descriptor of the update sites (.target file). Poms from the example
projects, require new dependencies, also, the versions of xtext and
Jbase.
The generator file was changed to print the aspects interceptions with
level "info"