Updated README and resolved problems when using several automatons
-First, it was updated the README
-Generator: When working with more than one automatons, the aspect
replicated the distribution of the message the same number of times than
automatons. This is, if you had n automatons, and execute one event,
this event was sent n times, different than just one time per event.
Also, changed some attributes names for avoiding conflicts
-Inferrer: Fixed the automaton definition, that alwways started the
automaton instead of save it in the 'instance' field. Finally, fixed a
bug that arises when it was declared two automatons that had the same
name for an state and a reaction was for each state, in this case, one
of the reactions was override by the other, so with this change, this
problem was solved.