Added a predecessor dependency.
Maven builds modules of the language according with some rules.
Sometimes, when any of their rules applied, last rule is select order
from parent's pom. However, it looks like there is some rules that are
applied and ignores parent's pom order. Therefore, to achieve the built
of the "lib" project first than the "lib.osgi" project, it has to be a
dependency of the first one into the second one to simulate precedence.