Improved priorities when calculating accuracy level
If in a label, one condition suggest an 'And' it must have a higher priority than a label that just accomplish a part, for example:

The property [](a->b) means that always after occurs 'a'  must occur 'b'.

S0=(TRUE> S1 |not(a)> S0), S1=(b-> S1 |not(a)&b > S0).

Transitions of S1 looks like they are redundant, however, they are not. If an 'a' occurs, in the next word (for maintain the prooperty) 'b' must be true; however, the cases are the following: {{b},{ab}}. If the case is the first one, automaton must move to S0, on the other hand, if both of them are true, it meant that 'a' was also true, so a 'b' have to appear in the next state, as the approach only consider unary events, in this case, 'b' was true, and must move to S0, and the label that represents this behaviour is followed by 'not(a)&b' instead of only 'a'.